janssen-guard-api/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
aiomysql==0.2.0
pymysql==1.1.0
alembic==1.12.1

# Validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
python-dotenv==1.0.0

